+++ src/lib.rs
@@ -454,3 +454,5 @@ mod uniffi_libtelio {
         }
     }
 }
+/// cbindgen:ignore
+pub use telio_starcast;